requirements-pytorch-dev.txt
mlflow
python-dotenv
torchvision>=0.15.1
torch>=2.0
lightning==2.0.0
jsonargparse[signatures]>=4.17.0
protobuf<4.0.0